requirements.txt
backports.tarfile==1.2.0
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
docutils==0.22
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.3.0
packaging==25.0
Pygments==2.19.2
pyproject_hooks==1.2.0
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
ruff==0.12.9
tomli==2.2.1
twine==6.1.0
urllib3==2.5.0
zipp==3.23.0
